seperated out files. most files could use some refactoring. todo: fix data being pulled in from dom in generateScale.js; todo: move css over SCSS
